+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -3,9 +3,9 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.repository.ReservationRepository;
+import roomescape.domain.reservation.ReservationRepository;
 import roomescape.domain.time.Time;
-import roomescape.domain.time.repository.TimeRepository;
+import roomescape.domain.time.TimeRepository;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
 import roomescape.dto.time.TimeRequest;
